[cleanup] moves routes folder into client
