+++ client/routes/routes.js
@@ -1,15 +1,15 @@
 import React from 'react';
 import { Router, Route, browserHistory, hashHistory, IndexRoute, useRouterHistory } from 'react-router';
 import { createHashHistory } from 'history';
-import App from '../client/components/App.js';
-import Singleplayer from '../client/containers/Singleplayer.js';
-import Multiplayer from '../client/containers/Multiplayer.js';
-import About from '../client/components/About';
-import LandingPage from '../client/components/LandingPage';
-import LandingPageMulti from '../client/containers/LandingPageMulti';
-import Login from '../client/components/Login';
-import Register from '../client/components/Register';
-import helperFunctions from '../client/utils/helperFunctions';
+import App from '../components/App.js';
+import Singleplayer from '../containers/Singleplayer.js';
+import Multiplayer from '../containers/Multiplayer.js';
+import About from '../components/About';
+import LandingPage from '../components/LandingPage';
+import LandingPageMulti from '../containers/LandingPageMulti';
+import Login from '../components/Login';
+import Register from '../components/Register';
+import helperFunctions from '../utils/helperFunctions';
 
 const appHistory = useRouterHistory(createHashHistory)({ queryKey: false});
